### 2024-10-15 4319a19
docs: minor fix for the paper
