+++ docs/paper/paper.bib
@@ -44,6 +44,22 @@
   urldate = {2023-11-15}
 }
 
+@article{kwon2024Equilipy,
+  title = {Equilipy: A Python Package for Calculating Phase Equilibria},
+  shorttitle = {Equilipy},
+  author = {Kwon, Sun Yong and Thibodeau, Eric and Plotkowski, Alex and Yang, Ying},
+  year = {2024},
+  month = aug,
+  journal = {Journal of Open Source Software},
+  volume = {9},
+  number = {100},
+  pages = {6875},
+  issn = {2475-9066},
+  doi = {10.21105/joss.06875},
+  urldate = {2024-10-14},
+  langid = {english}
+}
+
 @inproceedings{lam2015Numba,
   title = {Numba: A {{LLVM-based Python JIT}} Compiler},
   shorttitle = {Numba},
@@ -132,6 +148,22 @@
   langid = {english}
 }
 
+@article{pasquale2024CLEAVING,
+  title = {{{CLEAVING}}: A {{LAMMPS}} Package to Compute Surface Free Energies},
+  shorttitle = {{{CLEAVING}}},
+  author = {Pasquale, Nicodemo Di and Davidchack, Ruslan and Rovigatti, Lorenzo},
+  year = {2024},
+  month = feb,
+  journal = {Journal of Open Source Software},
+  volume = {9},
+  number = {94},
+  pages = {5886},
+  issn = {2475-9066},
+  doi = {10.21105/joss.05886},
+  urldate = {2024-10-14},
+  langid = {english}
+}
+
 @misc{qiang2024Scaling,
   title = {Scaling of Phase Count in Multicomponent Liquids},
   author = {Qiang, Yicheng and Luo, Chengjie and Zwicker, David},
@@ -187,6 +219,40 @@
   urldate = {2024-09-10}
 }
 
+@article{tse2022SurfinPy,
+  title = {{{SurfinPy}} 2.0: {{A Phase Diagram Generator}} for {{Surfaces}} and {{Bulk Phases}}},
+  shorttitle = {{{SurfinPy}} 2.0},
+  author = {Tse, Joshua S. and Molinari, Marco and Parker, Stephen C. and Symington, Adam R.},
+  year = {2022},
+  month = mar,
+  journal = {Journal of Open Source Software},
+  volume = {7},
+  number = {71},
+  pages = {4014},
+  issn = {2475-9066},
+  doi = {10.21105/joss.04014},
+  urldate = {2024-10-14},
+  langid = {english}
+}
+
+@article{virtanen2020SciPy,
+  title = {{{SciPy}} 1.0: Fundamental Algorithms for Scientific Computing in {{Python}}},
+  shorttitle = {{{SciPy}} 1.0},
[…]
