### 2024-10-15 #7304 @qiangyicheng
Thanks! We found several JOSS papers on related topics. So we would like to suggest the submitting authors of these papers. Here is the list:

`@jstse` from: Tse et al., (2022). SurfinPy 2.0: A Phase Diagram Generator for Surfaces and Bulk Phases. Journal of Open Source Software, 7(71), 4014, https://doi.org/10.21105/joss.04014
`@SunyongKwon` from: Kwon et al., (2024). Equilipy: a python package for calculating phase equilibria. Journal of Open Source Software, 9(100), 6875, https://doi.org/10.21105/joss.06875
`@demonico85` from: Di Pasquale et al., (2024). CLEAVING: a LAMMPS package to compute surface free energies. Journal of Open Source Software, 9(94), 5886, https://doi.org/10.21105/joss.05886
