### 2024-10-18 #7304 @srmnitc
@jstse @SunyongKwon @mastricker, would any of you have time and be willing to review this submission for JOSS? As you might know, we carry out our checklist-driven reviews here in GitHub issues and follow these [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). Please let me know if this would be interesting for you. We generally expect a review in 2-4 weeks. Thanks in advance!
