### 2024-10-21 #7388 @editorialbot
[REVIEW]: flory: A Python package for finding coexisting phases in multicomponent mixtures
**Submitting author:** @qiangyicheng (<a href="http://orcid.org/0000-0003-2053-079X">Yicheng Qiang</a>)
**Repository:** https://github.com/qiangyicheng/flory
**Branch with paper.md** (empty if default branch):
**Version:** 0.3.0
