### 2024-10-23 496896e
feat: allow zero fractions for `FreeEnergy`
