### 2024-10-28 qiangyicheng/flory#18 @SunyongKwon
[JOSS review] Add CONTRIBUTING.md file
Although "code_of_conduct.md" exists, contributing guidelines is rather unclear. Please, consider creating a `CONTRIBUTING.md` file (e.g. [https://contributing.md/example/](https://contributing.md/example/)).
