### 2024-10-28 qiangyicheng/flory#19 @SunyongKwon
[JOSS review] Dependency issue Matplotlib
A few examples require the `Matplotlib` package for visualization.
Please consider adding `Matplotlib` in the installation requirement, so that the users run examples without facing issues.
