### 2024-10-28 #7388 @SunyongKwon
Hi @srmnitc and @qiangyicheng, I have completed reviewing the the software and manuscript for `flory`.
### Overview
Overall, the software is very well organized. It was very easy to install and run examples. Documentation was clear and easy to follow. The workflow for packaging operates, and the test suite is implemented nicely.

Regarding the manuscript, the claims in `Statement of need` is a bit misleading. It is claimed that the benefit of using `flory` is the calculation efficiency as compared to other programs. However, relevant results have not been presented in the manuscript for such claims.

To my understanding, what differentiate `flory` from other minimizers is the capability of calculating multiphase-multicomponent equilibrium based on Flory-Huggins solution model. The Flory-Huggins model is not currently available in any CALPHAD-based programs; therefore, `flory` is essential for conducting such calculations.

One suggestion to improve the manuscript is providing the information regarding available Flory-Huggins Chi Database (for example, [3PDB](https://pppdb.uchicago.edu/chi)). Such information would be a great help for readers/users to seamlessly utilize `flory` for their research.

### Review summary
- (Documentation/Community guideline) Although "code_of_conduct.md" exists, clear contributing guidelines is missing. You may consider creating a `CONTRIBUTING.md` file (e.g.: [https://contributing.md/example/](https://contributing.md/example/)). [issue#18](https://github.com/qiangyicheng/flory/issues/18)
- (Example usage) Matplotlib should be added as a package for examples. [issue#19](https://github.com/qiangyicheng/flory/issues/19)
- (Software paper/State of the field) Consider revising `Statement of need` and providing the clear scope and high-level functionality of the program.
- (Minor review) Consider adding installation description in the documentation as well
