### 2024-11-13 #7388 @mastricker
Hi @qiangyicheng @srmnitc,

first of all, I would like to thank you for publishing your code! I
think this is very important and exposing oneself to review by an
outside party means more work but generally also means the code
quality and usability will be improved.

In general, the code looks good. Installation went smooth - I used a
conda environment. Examples run, tests run. I provide some comments
that I think will improve the overall experience.

## Manuscript
I am not very well versed with the actual computational problem but
generally from the field. To me, the manuscript seems like it is
written for an experienced person. I, personally, would benefit from
expanded section of the actual problem and why it is so
computationally demanding. Also the competitor software is generally
known to me (by name, not by usage) but your description and
comparison is lacking depth. The examples section would also benefit
from from a second example, maybe even a figure or a full
workflow. The "whole phase diagram" is mentioned, but not shown. To
you this might seem trivial, to me it is not. Also you mention
performance gains over competitors: can you provide numbers?

In addition, I also suggest that the paper would benefit from an
expanded "state of the field" section.

References appear fine.

## Code
Running the tests required additional installation of modules. Also
the code does not seem to follow a style convention (e.g. PEP8). When
I ran `flake8` (a commonly used linter), some wildcard imports were
found, which make it generally hard to check if imports are actually
used.
`black` (a commonly used autoformatter) essentially reformatted 33 of
53 source files which indicates that the code base is stylistically
not the same. Here is the output:

```
(flory) /flory$ black .
reformatted flory/docs/source/ext/autoclasstoc/errors.py
reformatted flory/examples/find_coexiting_phases.py
reformatted flory/examples/find_coexiting_phases_ternary.py
reformatted flory/examples/find_coexiting_phases_unequal_sizes.py
reformatted flory/examples/binary_phase_diagram.py
reformatted flory/docs/source/conf.py
reformatted flory/examples/finite_size_effect.py
reformatted flory/examples/grandcanonical_with_constraints.py
reformatted flory/flory/common/exceptions.py
reformatted flory/docs/source/ext/autoclasstoc/plugin.py
reformatted flory/examples/polydispersed.py
reformatted flory/flory/common/utilities.py
reformatted flory/flory/constraint/base.py
reformatted flory/flory/ensemble/__init__.py
reformatted flory/docs/source/ext/autoclasstoc/utils.py
reformatted flory/flory/entropy/__init__.py
reformatted flory/examples/ternary_phase_diagram.py
reformatted flory/flory/free_energy/__init__.py
reformatted flory/docs/source/ext/autoclasstoc/sections.py
reformatted flory/flory/common/phases.py
reformatted flory/flory/entropy/base.py
reformatted flory/flory/mcmp/__init__.py
reformatted flory/flory/constraint/linear_global.py
reformatted flory/flory/interaction/base.py
reformatted flory/flory/shortcut.py
reformatted flory/flory/constraint/linear_local.py
reformatted flory/tests/ensemble/test_canonical.py
reformatted flory/flory/entropy/ideal_gas.py
reformatted flory/tests/mcmp/test_finder.py
reformatted flory/tests/test_examples.py
reformatted flory/tests/test_consistency.py
reformatted flory/flory/mcmp/_finder_impl.py
reformatted flory/flory/mcmp/finder.py

All done! ✨ 🍰 ✨
33 files reformatted, 20 files left unchanged.
```
Please check for stylistic consistency. This also makes it easier for
contributors to add code in a coherent/consistent style. Or maybe you
followed a style but it is not defined somewhere.

For running the tests, I additionally needed to install `pytest`,
`pytest-cov`, and `matplotlib`. Some `test_examples` are also skipped by
default. Maybe these should be run?
