### 2024-11-20 qiangyicheng/flory#20 @srmnitc
Changes to manuscript
Part of review over at https://github.com/openjournals/joss-reviews/issues/7388

#### Points raised by @mastricker

I am not very well versed with the actual computational problem but
generally from the field. To me, the manuscript seems like it is
written for an experienced person. I, personally, would benefit from
expanded section of the actual problem and why it is so
computationally demanding. Also the competitor software is generally
known to me (by name, not by usage) but your description and
comparison is lacking depth. The examples section would also benefit
from from a second example, maybe even a figure or a full
workflow. The "whole phase diagram" is mentioned, but not shown. To
you this might seem trivial, to me it is not. Also you mention
performance gains over competitors: can you provide numbers?

In addition, I also suggest that the paper would benefit from an
expanded "state of the field" section.

#### Points raised by @SunyongKwon

Regarding the manuscript, the claims in Statement of need is a bit misleading. It is claimed that the benefit of using flory is the calculation efficiency as compared to other programs. However, relevant results have not been presented in the manuscript for such claims.

To my understanding, what differentiate flory from other minimizers is the capability of calculating multiphase-multicomponent equilibrium based on Flory-Huggins solution model. The Flory-Huggins model is not currently available in any CALPHAD-based programs; therefore, flory is essential for conducting such calculations.

One suggestion to improve the manuscript is providing the information regarding available Flory-Huggins Chi Database (for example, [3PDB](https://pppdb.uchicago.edu/chi)). Such information would be a great help for readers/users to seamlessly utilize flory for their research.
