### 2024-11-20 qiangyicheng/flory#21 @srmnitc
Style changes in code
Part of review over at https://github.com/openjournals/joss-reviews/issues/7388

Point raised by @mastricker

Running the tests required additional installation of modules. Also
the code does not seem to follow a style convention (e.g. PEP8). When
I ran flake8 (a commonly used linter), some wildcard imports were
found, which make it generally hard to check if imports are actually
used.
black (a commonly used autoformatter) essentially reformatted 33 of
53 source files which indicates that the code base is stylistically
not the same.

Please check for stylistic consistency. This also makes it easier for
contributors to add code in a coherent/consistent style. Or maybe you
followed a style but it is not defined somewhere.
