### 2024-11-20 qiangyicheng/flory#22 @srmnitc
Additional modules needed for tests
Part of review over at https://github.com/openjournals/joss-reviews/issues/7388

Point raised by @mastricker

For running the tests, I additionally needed to install pytest,
pytest-cov, and matplotlib. Some test_examples are also skipped by
default. Maybe these should be run?
