### 2024-11-20 qiangyicheng/flory#23 @srmnitc
Add installation description in documentation
Part of review over at https://github.com/openjournals/joss-reviews/issues/7388

Points raised by @SunyongKwon

Consider adding installation description in the documentation as well
