### 2024-12-01 e2dbdd2
docs: add CONTRIBUTING.md
