### 2024-12-01 qiangyicheng/flory#18 @qiangyicheng
Thank you for the comment and the suggestion for the `CONTRIBUTING.md` generator. We have added the `CONTRIBUTING.md` file. We choose to keep `CONTRIBUTING.md` brief and short to ease any contribution to the package.
