+++ flory/free_energy/__init__.py
@@ -3,5 +3,6 @@
 .. codeauthor:: Yicheng Qiang <yicheng.qiang@ds.mpg.de>
 .. codeauthor:: David Zwicker <david.zwicker@ds.mpg.de>
 """
+
 from .base import FreeEnergyBase
 from .flory_huggins import FloryHuggins
