+++ flory/entropy/__init__.py
@@ -2,6 +2,7 @@
 
 .. codeauthor:: Yicheng Qiang <yicheng.qiang@ds.mpg.de>
 """
+
 from .base import EntropyBase, EntropyBaseCompiled
 from .ideal_gas import IdealGasEntropy, IdealGasEntropyCompiled
 from .ideal_gas_polydispersed import (
