+++ docs/source/ext/autoclasstoc/errors.py
@@ -1,9 +1,8 @@
 class AutoClassTocError(Exception):
     pass
 
+
 class ConfigError(AutoClassTocError):
     """
     Indicate an configuration error affecting :rst:dir:`autoclasstoc`.
     """
-
-
