+++ flory/constraint/base.py
@@ -10,7 +10,7 @@ from ..common import filter_kwargs
 
 class ConstraintBaseCompiled:
     r"""Abstract base class for a general compiled constraint.
-    
+
     This abstract class defines the necessary members of a compiled constraint instance.
     This abstract class does not inherit from :class:`abc.ABC`, since the
     :func:`numba.experimental.jitclass` currently does not support some members of
@@ -30,7 +30,7 @@ class ConstraintBaseCompiled:
     @property
     def potential(self) -> np.ndarray:
         r"""The potential for features generated by the constraint.
-        
+
         This property typically contains the Jacobian of the constraint part of the free
         energy with respect to the volume fractions of the features, which is an array
         with the size of :math:`N_\mathrm{S} \times N_\mathrm{M}`. Note that this property should be
@@ -41,7 +41,7 @@ class ConstraintBaseCompiled:
     @property
     def volume_derivative(self) -> np.ndarray:
         r"""The volume derivatives of the constraint part of entropic energy.
-        
+
         This property typically contains the Jacobian of the constraint part of the free
         energy with respect to the volumes of the compartments, which is an array with the
         size of :math:`N_\mathrm{M}`. Note that this property should be used after :meth:`prepare` is
@@ -91,7 +91,7 @@ class ConstraintBaseCompiled:
             masks:
                 Constant. The 1D array with the size of :math:`N_\mathrm{M}`, containing the masks to
                 mark whether the compartment is living or not.
-                
+
         Returns:
             : the max absolute residue.
         """
