+++ flory/mcmp/__init__.py
@@ -1,3 +1,3 @@
-"""Module containing finder for coexisting phases.
-"""
+"""Module containing finder for coexisting phases."""
+
 from .finder import CoexistingPhasesFinder
