+++ examples/finite_size_effect.py
@@ -17,9 +17,9 @@ phi_means = np.full(N_comp, 1.0 / N_comp)  # set a symmetric composition
 
 line_N_compartment = []
 line_N_phase = []
-for N_compartment in range(4, 16, 2): # use different compartment number
+for N_compartment in range(4, 16, 2):  # use different compartment number
     phases = flory.find_coexisting_phases(
-        N_comp, chis, phi_means, num_part = N_compartment, progress=True
+        N_comp, chis, phi_means, num_part=N_compartment, progress=True
     )
     line_N_compartment.append(N_compartment)
     line_N_phase.append(phases.volumes.shape[0])

[…]
