### 2024-12-01 5504039
ops: provide `doc` and `test` install options
