+++ pyproject.toml
@@ -29,6 +29,10 @@ dependencies = { file = ["requirements.txt"] }
 optional-dependencies = { dev = { file = [
     "tests/requirements.txt",
     "docs/requirements.txt",
+] }, test = { file = [
+    "tests/requirements.txt",
+] }, doc = { file = [
+    "docs/requirements.txt",
 ] } }
 
 [tool.setuptools_scm]
@@ -67,27 +71,46 @@ docstring-code-format = true
 [tool.ruff.lint]
 select = [
     "UP",  # pyupgrade
-    "I",  # isort
-    "A",  # flake8-builtins 
-    "B",  # flake8-bugbear
+    "I",   # isort
+    "A",   # flake8-builtins 
+    "B",   # flake8-bugbear
     "C4",  # flake8-comprehensions
-    "FA", # flake8-future-annotations
+    "FA",  # flake8-future-annotations
     "ISC", # flake8-implicit-str-concat
     "ICN", # flake8-import-conventions
     "LOG", # flake8-logging
-    "G", # flake8-logging-format
+    "G",   # flake8-logging-format
     "PIE", # flake8-pie
-    "PT", # flake8-pytest-style
-    "Q", # flake8-quotes
+    "PT",  # flake8-pytest-style
+    "Q",   # flake8-quotes
     "RSE", # flake8-raise
     "RET", # flake8-return
     "SIM", # flake8-simplify
     "PTH", # flake8-use-pathlib
 ]
-ignore = ["B007", "B027", "B028", "SIM108", "ISC001", "PT006", "PT011", "PTH123", "RET504", "RET505", "RET506"]
+ignore = [
+    "B007",
+    "B027",
+    "B028",
+    "SIM108",
+    "ISC001",
+    "PT006",
+    "PT011",
+    "PTH123",
+    "RET504",
+    "RET505",
+    "RET506",
+]
 
 [tool.ruff.lint.isort]
-section-order = ["future", "standard-library", "third-party", "first-party", "self", "local-folder"]
+section-order = [
+    "future",
+    "standard-library",
+    "third-party",
+    "first-party",
+    "self",
+    "local-folder",
+]
 
 [tool.ruff.lint.isort.sections]
 self = ["flory"]
