### 2024-12-01 925bca4
docs: add info for optional dependencies
