### 2024-12-01 qiangyicheng/flory#21 @qiangyicheng
Thank you for pointing out the stylistic inconsistencies. We have rerun the formatter and also made it clear in [CONTRIBUTING.md](https://github.com/qiangyicheng/flory/blob/main/CONTRIBUTING.md) that we use `ruff` as the formatter.
