### 2024-12-01 3aaea79
ops: add `example` optional dependencies
