### 2024-12-01 qiangyicheng/flory#22 @qiangyicheng
Thank you for testing the package. We treat these packages as optional dependencies since they are not referenced by the main functionality of our package. Therefore, we chose not to include them directly and keep the dependencies of the package minimal. We provided an installation option `dev` to install these optional dependencies with `pip`, but did not make it clear in the documentation. Therefore, we have improved the documentation, and provided options `example`, `test` and `doc` to choose among groups of optional dependencies. Regarding the tests, we skipped some slow tests by default. These tests can be enabled with the `--runslow` option. We have also explained this in the revised `README.md`.
