### 2024-12-01 qiangyicheng/flory#19 @qiangyicheng
Thank you for the helpful suggestion.  We did not include `Matplotlib` since it is bulky and not directly referenced by the main functionality of our package. Therefore we chose to treat it as an optional dependency that can be installed by a `dev` option. We have revised `README.md` to clarify this, and provided options `example`, `test` and `doc` to choose among different groups of optional dependencies.
