### 2024-12-01 ab5d6d9
docs: add installation
