### 2024-12-01 40e30bf
docs: minor fixes
