+++ docs/source/installation.rst
@@ -1,4 +1,4 @@
-installation
+Installation
 ============
 
 Install using pip
