### 2024-12-01 qiangyicheng/flory#23 @qiangyicheng
Thank you for the suggestion. We have added the installation page accordingly.
