+++ docs/paper/paper.md
@@ -15,7 +15,7 @@ authors:
 affiliations:
  - name: Max Planck Institute for Dynamics and Self-Organization, Göttingen, Germany
    index: 1
-date: 9 September 2024
+date: 2 December 2024
 bibliography: paper.bib
 ---
 
@@ -23,21 +23,33 @@ bibliography: paper.bib
 
 Phase separation is an intrinsic property of mixtures that is widely observed in many scenarios, ranging from the simple demixing of oil and water to the condensation of biomolecules in cells.
 In multicomponent mixtures, phase separation can lead to many coexisting phases, which is crucial in many fields.
+One key step to understand phase separation is to measure or predict the composition of the coexisting phases.
 To support such research, the `flory` package provides an easily accessible, performant, and extensible code that finds coexisting phases in multicomponent mixtures.
 The package expresses the free energy of the mixtures by several orthogonal aspects to cover a broad range of physical situations.
 In contrast to existing methods, the `flory` package implements a state-of-art method that is optimized for mixtures of many components.
+The package mainly focus on the mixtures with uniform and relatively simple free energies such as Flory-Huggins free energy and its generalizations, while more complicated ones can also be supported through extensions.
 
 # Statement of need
 
-Finding coexisting phases is a general task in many fields, such as chemical engineering [@lukas2007Computational] and soft matter physics [@jacobs2023Theory].
-There are several strategies to theoretically predict coexisting phases [@jacobs2023Theory], including direct spatially-resolved simulations [@shrinivas2021Phase], the construction of the convex hull of the free energy  [@mao2019Phase], solving the balance equations [@zwicker2022Evolved], and direct minimization of the free energy [@lukas2007Computational].
-There are a few open-source packages that implement these strategies [@kwon2024Equilipy],.
+Finding coexisting phases is a common task in many fields, such as chemical engineering [@lukas2007Computational] and soft matter physics [@jacobs2023Theory].
+In general, the coexisting phases can be theoretically predicted by solving the balance equations between phases [@zwicker2022Evolved], or equivalently minimizing the total free energy of the whole mixture [@lukas2007Computational].
+Other strategies include direct spatially-resolved simulations [@shrinivas2021Phase] and the construction of the convex hull of the free energy landscape [@mao2019Phase].
+There are a few open-source packages that implement these strategies.
 Most notably, Calphad packages, including `Equilipy` [@kwon2024Equilipy], `pycalphad` [@otis2017Pycalphad] and `OpenCalphad` [@sundman2015Implementation], combine a database of candidate phases and several strategies above to compute phase diagrams of mixtures with few components.
 In addition, `SurfinPy` [@tse2022SurfinPy] applies the free energy minimization strategy to surface phases.
-However, the computational cost of finding the coexisting phases scales strongly with component count, so systems with many different components cannot be analyzed efficiently with these packages.
 
-The `flory` Python package is designed to fill this gap.
-It determines coexisting phases in a broad range of multicomponent mixtures while also being efficient enough for sampling entire phase diagrams when the composition is varied for many components.
+Despite of the multiple useful strategies and packages above, finding coexisting phases is still challenging when the number of the components $N_\mathrm{C}$ in the mixture is large.
+The reason is that with larger $N_\mathrm{C}$, the degree of freedom increases (e.g. composition of the phases), leading to the problem of global minimization or sampling in high-dimensional space.
+For example, the convex hull strategy needs to sample the entire free energy landscape.
[…]
