### 2024-12-02 a7a3ba7
Revised paper
