### 2024-12-05 qiangyicheng/flory#20 @qiangyicheng
We thank both reviewers for the helpful comments.

Both reviewers raised questions about the comparison between our package and previous ones. Indeed our original text in the “Statement of need” section was a bit misleading and did not stress the purpose of our package sufficiently: Compared with previous packages (especially CALPHAD ones), which focused on the various forms of free energy from databases, our package focuses on mixtures with many components described by simple free energies, e.g. a Flory-Huggins one. The main task of our package is thus more theoretical aimed at general multicomponent mixtures. This is also the reason why we did not include a direct comparison of the performance between our package and previous ones. We have revised our manuscript to clarify this.

Concerning the other points raised by reviewer @mastricker:
Regarding the actual computational problem, we now explain it more clearly in the manuscript. Essentially, it comes from the problem of optimisation and sampling in high-dimensional space. We also include a new example for generating a sample phase diagram in the manuscript. We suspect the reviewer meant the “Statement of need” section when referring to “state of the field”, which we have revised and expanded to address the points above.

Concerning the other points raised by reviewer @SunyongKwon:
We thank the reviewer for suggesting the 3PDB database. We have added it in the revised manuscript.
