### 2024-12-05 #7388 @qiangyicheng
We thank the editor and both reviewers again for investing time on our manuscript and the package. Both reviewers raised points on the comparison between our package and existing ones. We have revised our manuscript to clarify that our package has a different focus compared to previous packages. Regarding the dependency issues, we now keep the default dependencies simple, and treat others as optional dependencies. We have improved the documentation to make this clear and provided a more detailed installation guide. We have also responded to all other points in the corresponding Github issues.
