### 2024-12-10 #7388 @srmnitc
> @editorialbot generate pdf

This is the right way to do it :)
