### 2024-12-10 #7388 @SunyongKwon
@qiangyicheng
I would suggest the authors to revise two sentences in the manuscript: 'While the existing Calphad packages implement some algorithms that are more suitable, the packages were optimized for systems with few components and instead provide a high flexibility on the free energies of candidate phases. Since there are currently no packages focusing on the many-component case to our knowledge'

For the authors' information, CALPHAD packages can calculate a multicomponent system (e.g. over 25 elements with over 1000 different species) within a reasonable time window (less than a minute in a desktop). Therefore, 'the packages were optimized for systems with few components' is simply not true.

In my opinion, the following sentence 'Compared with the existing Calphad packages, flory focuses on the case where all candidate phases share the same free energy function, e.g., the simple Flory-Huggins free energy.` serves the purpose for the statement of the need.

@srmnitc Thank you for reminding me for double checking. I confirm here that I recommend the publication, once the authors revise the manuscript.
