### 2024-12-11 c1f94b2
docs: update manuscript
