+++ docs/paper/paper.bib
@@ -76,6 +76,19 @@
   langid = {english}
 }
 
+@article{li2020Implementation,
+  title = {Implementation of the {{UNIQUAC}} Model in the {{OpenCalphad}} Software},
+  author = {Li, Jing and Sundman, Bo and Winkelman, Jozef G. M. and Vakis, Antonis I. and Picchioni, Francesco},
+  year = {2020},
+  month = mar,
+  journal = {Fluid Phase Equilibria},
+  volume = {507},
+  pages = {112398},
+  issn = {0378-3812},
+  doi = {10.1016/j.fluid.2019.112398},
+  urldate = {2024-12-11}
+}
+
 @book{lukas2007Computational,
   title = {Computational Thermodynamics: The {{Calphad}} Method},
   shorttitle = {Computational Thermodynamics},
