### 2024-12-12 52cb557
docs: refine explanations and improve clarity in paper
