### 2024-12-12 c69c093
docs: update bibliography and enhance manuscript clarity
