+++ docs/paper/paper.bib
@@ -180,7 +180,9 @@
 @misc{Polymera,
   title = {Polymer {{Property Predictor}} and {{Database}}},
   urldate = {2024-12-02},
-  howpublished = {https://pppdb.uchicago.edu/}
+  year = 2024,
+  howpublished = {https://pppdb.uchicago.edu/},
+  note = {Accessed: 2024-12-10}
 }
 
 @misc{qiang2024Scaling,
