+++ docs/paper/paper.md
@@ -43,15 +43,9 @@ This is because the number of degrees of freedom (e.g., to describe the composit
 Moreover, the possible number of coexisting phases also increases with $N_\mathrm{C}$ according to Gibbs phase rule, implying that the free energy of the entire system comprises roughly $N_\mathrm{C}^2$ free variables.
 This high-dimensional space needs to be sampled to find the global minimum with multiple coexisting phases, which is infeasible for some of the strategies mentioned above since they become prohibitively expensive.
 For example, the cost of the convex hull strategy increases exponentially with $N_\mathrm{C}$ since it requires to sample the entire free energy landscape.
-The existing Calphad packages address this challenge by implementing several strategies and providing high flexibility on candidate phases, allowing each phase to have different free energies.
-Compared with existing Calphad packages, `flory` focuses on the case where all candidate phases share the same free energy function, e.g., the simple Flory-Huggins free energy.
-Similar concept was used in a recent submodule of the `OpenCalphad` package [@li2020Implementation].
-WE NEED TO STRESS THAT OUR ASSUMPTIONS ALLOW FOR A MORE EFFICIENT IMPLEMENTATION (if this is true)
-These simple models are more common in systems such as polymeric liquid mixtures, and considered relevant in the phase separation in biology recently. 
-The value of the Flory-Huggins parameters can either be obtained from database such as `3PDB` [@Polymera] for realistic polymer mixtures, or freely chosen for theoretical investigations.
-Using these assumptions, the algorithm implemented by the package makes use of the physical information of the free energy to explore the high-dimensional free energy landscape efficiently.
-THE NEXT SENTENCE IS AN IMPLEMENTATION DETAIL. I THINK WE RATHER WANT TO DESCRIBE WHAT THE GOAL OF THE PACKAGE IS
-To obtain the coexisting phases in equilibrium without the prior knowledge of the compositions of the phases, the `flory` package usually starts from many phases initially and cluster the equivalent phases afterwards.
+The existing Calphad packages address this challenge by implementing several strategies and providing high flexibility on candidate phases, allowing each phase to have different free energies to model realistic systems.
+In contrast, `flory` focuses on the general physics of multicomponent phase separation, and thus assumes that all candidate phases share the same free energy function, e.g., the simple Flory-Huggins free energy, similar to a recent submodule of the `OpenCalphad` package [@li2020Implementation].
+The restrictions of the physical model simplifies the user interface and allow for an efficient optimization algorithm.
 As the result, the `flory` package can determine the multiple coexisting phases in equilibrium in a range of multicomponent mixtures while also being efficient enough when the number of the components $N_\mathrm{C}$ is large.
 
 # Methods
@@ -65,6 +59,7 @@ $$\bar{f}({N_\mathrm{P}}, \{J_p\}, \{\phi_{p,i}\}) = \sum_{p=1}^{{N_\mathrm{P}}}
 where $N_\mathrm{C}$ is the number of components, $N_\mathrm{P}$ is the number of phases, $J_p$ denotes the fraction of volume that phase $p=1,\ldots,N_\mathrm{P}$ occupies in the entire system, and $\phi_{p,i}$ is the volume fraction of component $i=1,\ldots,N_\mathrm{C}$ in phase $p$.
 The physical behavior of the mixture is encoded in the free energy density $f$, which `flory` expresses using four orthogonal aspects: interaction, entropy, ensemble, and constraints.
 The package only imposes limits on the entropy part, which is crucial for the core algorithm, while the other three aspects are rather flexible.
[…]
