### 2024-12-12 #7388 @qiangyicheng
@SunyongKwon Thank you for the suggestion and the information on the Calphad method. We have revised the manuscript. Besides, we also notice a recent submodule of OpenCalphad that uses concept similar to ours, which we have added to our manuscript as well. Regarding the Calphad calculation with many components, could you please suggest some reference that we can look into? We suspect there's some difference between the use case of our package and the mentioned Calphad one, since It seems to be extreme challenging to conduct global optimization for a function with 1000 variables. Maybe this could help us further improve our manuscript and package.
