### 2024-12-12 #7388 @SunyongKwon
@qiangyicheng
There are many literature for minimizing Gibbs energies in CALPHAD communities. Most of the algorithms are based on a Lagrangian multiplier method proposed by [White et al.](https://doi.org/10.1063/1.1744264) in 1958. Some notable progress for this method is by Eriksson(G. Eriksson, Acta Chem. Scand. 1971) and [Hillert](https://www-sciencedirect-com.ornl.idm.oclc.org/science/article/pii/0378436381910007). OpenCalphad is based on the method suggested by Hillert and further refined by [Sundman et al.](https://www-sciencedirect-com.ornl.idm.oclc.org/science/article/pii/S0927025615000361)

FYI, the CALPHAD method involves minimization of multiple Gibbs energy functions where each function can have more than 1000 species. As an example, a commercial CALPHAD program callled FactSage (based on the algorithm developed by Eriksson) can calculate global minimum of 1500 Gibbs energies with 7000 species in total.

I do not anticipate a complete literature review in your manuscript. However, it is worth to note that there are efficient algorithms for free energy minimization for multicomponent-multiphase minimizations.
