### 2025-01-15 51b27af
docs: revise manuscript
