+++ docs/paper/paper.md
@@ -35,24 +35,25 @@ Finding coexisting phases is a common task in many fields, such as chemical engi
 In general, the coexisting phases can be theoretically predicted by solving the balance equations between phases [@zwicker2022Evolved], or equivalently minimizing the total free energy of the whole mixture [@lukas2007Computational].
 Other strategies include direct spatially-resolved simulations [@shrinivas2021Phase] and the construction of the convex hull of the free energy landscape [@mao2019Phase].
 There are a few open-source packages that implement these strategies.
-Most notably, Calphad packages, including `Equilipy` [@kwon2024Equilipy], `pycalphad` [@otis2017Pycalphad] and `OpenCalphad` [@sundman2015Implementation], combine a database of candidate phases and several strategies above to compute phase diagrams of mixtures with few components.
+Most notably, Calphad packages, including `Equilipy` [@kwon2024Equilipy], `pycalphad` [@otis2017Pycalphad] and `OpenCalphad` [@sundman2015OpenCalphad], combine a database of candidate phases and several strategies above to compute phase diagrams of mixtures with few components.
 In addition, `SurfinPy` [@tse2022SurfinPy] applies the free energy minimization strategy to surface phases.
 
-Despite these useful strategies and packages, finding coexisting phases is still challenging for mixtures with a large number of the components, $N_\mathrm{C}$.
+Generally, finding coexisting phases is challenging for mixtures with a large number of the components, $N_\mathrm{C}$.
 This is because the number of degrees of freedom (e.g., to describe the composition of the phases) increases with larger $N_\mathrm{C}$.
 The possible number of coexisting phases also increases according the Gibbs phase law, leading to a free energy function of approximately $N_\mathrm{C}^2$ free variables.
 Finding the multiple coexisting phases then requires a global minimization and sampling in such a high-dimensional space.
 Consequently, some of the strategies mentioned above become prohibitively expensive.
 For example, the cost of the convex hull strategy increases exponentially with $N_\mathrm{C}$ since it requires to sample the entire free energy landscape.
-The existing Calphad packages are dedicated to address this challenge by implementing several strategies. 
+The existing Calphad packages are dedicated to address this challenge by combining several strategies. 
 Specifically, Calphad packages usually provide high flexibility on candidate phases, allowing each phase to have different free energies.
 Compared with the existing Calphad packages, `flory` focuses on the case where all candidate phases share the same free energy function, e.g., the simple Flory-Huggins free energy.
 Similar concept was used in a recent submodule of the `OpenCalphad` package [@li2020Implementation].
 These simple models are more common in systems such as polymeric liquid mixtures, and considered relevant in the phase separation in biology recently. 
 The value of the Flory-Huggins parameters can either be obtained from database such as `3PDB` [@Polymera] for realistic polymer mixtures, or freely chosen for theoretical investigations.
-Using these assumptions, the algorithm implemented by the package makes use of the physical information of the free energy to explore the high-dimensional free energy landscape efficiently.
-To obtain the coexisting phases in equilibrium without the prior knowledge of the compositions of the phases, `flory` package usually starts from many phases initially and cluster the equivalent phases afterwards.
-As the result, the `flory` package can determine the multiple coexisting phases in equilibrium in a range of multicomponent mixtures while also being efficient enough when the number of the components $N_\mathrm{C}$ is large.
+With these simple models, `flory` removes some explicit constraints in the numerics, thus reducing the Lagrange multipliers required [@white1958Chemical].
+The algorithm implemented by the package makes use of the physical information of the free energy to explore the high-dimensional free energy landscape efficiently.
[…]
