+++ docs/paper/paper.bib
@@ -29,6 +29,21 @@
   langid = {english}
 }
 
+@article{hillert1981Discussion,
+  title = {A Discussion of Methods of Calculating Phase Diagrams},
+  author = {Hillert, Mats},
+  year = {1981},
+  month = dec,
+  journal = {Bulletin of Alloy Phase Diagrams},
+  volume = {2},
+  number = {3},
+  pages = {265--268},
+  issn = {0197-0216},
+  doi = {10.1007/BF02868269},
+  urldate = {2025-01-15},
+  langid = {english}
+}
+
 @article{jacobs2023Theory,
   title = {Theory and {{Simulation}} of {{Multiphase Coexistence}} in {{Biomolecular Mixtures}}},
   author = {Jacobs, William M.},
@@ -238,6 +253,21 @@
   urldate = {2024-09-10}
 }
 
+@article{sundman2015OpenCalphad,
+  title = {{{OpenCalphad}} - a Free Thermodynamic Software},
+  author = {Sundman, Bo and Kattner, Ursula R. and Palumbo, Mauro and Fries, Suzana G.},
+  year = {2015},
+  month = dec,
+  journal = {Integr Mater Manuf Innov},
+  volume = {4},
+  number = {1},
+  pages = {1--15},
+  issn = {2193-9772},
+  doi = {10.1186/s40192-014-0029-1},
+  urldate = {2025-01-15},
+  langid = {english}
+}
+
 @article{tse2022SurfinPy,
   title = {{{SurfinPy}} 2.0: {{A Phase Diagram Generator}} for {{Surfaces}} and {{Bulk Phases}}},
   shorttitle = {{{SurfinPy}} 2.0},
@@ -270,6 +300,20 @@
   urldate = {2024-10-15},
   copyright = {2020 The Author(s)},
   langid = {english}
+}
+
+@article{white1958Chemical,
+  title = {Chemical {{Equilibrium}} in {{Complex Mixtures}}},
+  author = {White, W. B. and Johnson, S. M. and Dantzig, G. B.},
+  year = {1958},
+  month = may,
+  journal = {The Journal of Chemical Physics},
+  volume = {28},
+  number = {5},
+  pages = {751--755},
+  issn = {0021-9606},
+  doi = {10.1063/1.1744264},
+  urldate = {2025-01-15}
 }
 
 @article{zwicker2022Evolved,
