+++ docs/paper/paper.md
@@ -38,22 +38,22 @@ There are a few open-source packages that implement these strategies.
 Most notably, Calphad packages, including `Equilipy` [@kwon2024Equilipy], `pycalphad` [@otis2017Pycalphad] and `OpenCalphad` [@sundman2015OpenCalphad], combine a database of candidate phases and several strategies above to compute phase diagrams of mixtures with few components.
 In addition, `SurfinPy` [@tse2022SurfinPy] applies the free energy minimization strategy to surface phases.
 
-Generally, finding coexisting phases is challenging for mixtures with a large number of the components, $N_\mathrm{C}$.
+In general, finding coexisting phases is challenging for mixtures with a large number of the components, $N_\mathrm{C}$.
 This is because the number of degrees of freedom (e.g., to describe the composition of the phases) increases with larger $N_\mathrm{C}$.
-The possible number of coexisting phases also increases according the Gibbs phase law, leading to a free energy function of approximately $N_\mathrm{C}^2$ free variables.
+The possible number of coexisting phases also increases according the Gibbs phase law, leading to a free energy function with approximately $N_\mathrm{C}^2$ free variables.
 Finding the multiple coexisting phases then requires a global minimization and sampling in such a high-dimensional space.
 Consequently, some of the strategies mentioned above become prohibitively expensive.
-For example, the cost of the convex hull strategy increases exponentially with $N_\mathrm{C}$ since it requires to sample the entire free energy landscape.
-The existing Calphad packages are dedicated to address this challenge by combining several strategies. 
-Besides, Calphad packages usually provide high flexibility on candidate phases, allowing each phase to have different free energies.
-Compared with the existing Calphad packages, `flory` focuses on the case where all candidate phases share the same free energy function, e.g., the simple Flory-Huggins free energy.
-These simple models are more common in liquid systems such as polymeric mixtures, and considered relevant in the phase separation in biological cells recently. 
-Similar concept was also used in a recent submodule of the `OpenCalphad` package [@li2020Implementation].
-The `flory` package does not contain a database for interaction parameters.
+For example, the cost of the convex hull strategy increases exponentially with $N_\mathrm{C}$ since it requires sampling the entire free energy landscape.
+Existing Calphad packages are designed to address this challenge by combining multiple strategies. 
+Moreover, Calphad packages usually provide high flexibility on candidate phases, allowing each phase to have different free energies.
+Compared to the existing Calphad packages, `flory` focuses on the case where all candidate phases share the same free energy function, e.g. the simple Flory-Huggins free energy.
+These simple models are more common in liquid systems such as polymer mixtures, and have recently been considered relevant to phase separation in biological cells. 
+A similar concept has been used in a recent submodule of the `OpenCalphad` package [@li2020Implementation].
+The `flory` package does not contain a database of interaction parameters.
 The value of the interaction parameters can either be obtained from database such as `3PDB` [@Polymera] for realistic polymer mixtures, or freely chosen for theoretical investigations.
-With these simple models, `flory` removes some explicit constraints in the numerics, thus reducing the Lagrange multipliers required [@white1958Chemical].
-The algorithm implemented by the package makes use of the physical information of the free energy to explore the high-dimensional free energy landscape efficiently.
-To obtain the coexisting phases in equilibrium without the prior knowledge of the compositions of the phases, `flory` package starts from many phases initially and cluster the equivalent phases afterwards.
+Focusing on these simple models, `flory` removes some explicit constraints in the numerics, thus reducing the Lagrange multipliers required [@white1958Chemical].
+The algorithm implemented by the package uses the physical information of the free energy to explore the high-dimensional free energy landscape efficiently.
[…]
