### 2025-01-15 a5c5447
docs: simplify the text a bit
