### 2025-01-15 #7388 @qiangyicheng
Thanks for the reminder @srmnitc. We have just revised the manuscript to address these comments. In particular, we now make it clear that existing CALPHAD packages provide powerful multiphase coexistence solvers with high flexibility. In contrast, our package focuses more on simple free energies, such as Flory-Huggins free energies, allowing for more simplification in both usage and numerics.
