### 2025-02-11 #7388 @qiangyicheng
@srmnitc Thanks a lot! I have double checked the metadata and achieve the new release:

- The version number: 0.3.0
- The DOI of archive: 10.5281/zenodo.14851873
