### 2025-02-12 6d8697f
Limited maximal python version and added checks for python 3.13
