+++ pyproject.toml
@@ -8,7 +8,7 @@ description = "Python package for finding coexisting phases in multicomponent mi
 authors = [{ name = "Yicheng Qiang", email = "yicheng.qiang@ds.mpg.de" }]
 license = { text = "MIT License" }
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.9,<3.14"
 dynamic = ["version", "dependencies", "optional-dependencies"]
 
 keywords = ["physics", "phase-separation", "free-energy"]
@@ -22,6 +22,7 @@ classifiers = [
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
 ]
 
 [tool.setuptools.dynamic]
