### 2025-02-17 qiangyicheng/flory#32 @srmnitc
Changes to paper
Part of review over at https://github.com/openjournals/joss-reviews/issues/7388

Here are some final comments on the paper from my side:

- line 9: a citation could strengthen this line
- line 39: could you add a citation that clarifies what these several strategies are?
- line 41: a citation/at least a link to calphad tools that documents this
- line 109: maybe a link to the docs could be helpful
- references: please use either full journal names or acronyms consistently

Please feel free to close this issue once you implement the changes. We will the finish the rest of the steps. Thanks!
