### 2025-02-17 #7388 @srmnitc
@qiangyicheng I opened an issue with some minor comments about the paper. Could you please take a look? We can finish up here as soon as the issue is closed. Thanks again for this very interesting package!
