### 2025-02-18 c5c3d38
docs: revise paper according to editor's comments
