+++ docs/paper/paper.bib
@@ -1,3 +1,17 @@
+@misc{2019Polymer,
+  title = {Polymer {{Property Predictor}} and {{Database}}},
+  year = {2019},
+  url = {https://pppdb.uchicago.edu},
+  urldate = {2024-12-02}
+}
+
+@misc{2025Flory,
+  title = {`Flory' {{Python}} Package --- Flory Documentation},
+  year = {2025},
+  url = {https://flory.readthedocs.io},
+  urldate = {2025-02-18}
+}
+
 @article{andersson2002ThermoCalc,
   title = {Thermo-{{Calc}} \& {{DICTRA}}, Computational Tools for Materials Science},
   author = {Andersson, J-O and Helander, Thomas and H{\"o}glund, Lars and Shi, Pingfang and Sundman, Bo},
@@ -17,7 +31,7 @@
   author = {George, Easo P. and Raabe, Dierk and Ritchie, Robert O.},
   year = {2019},
   month = aug,
-  journal = {Nat Rev Mater},
+  journal = {Nature Reviews Materials},
   volume = {4},
   number = {8},
   pages = {515--534},
@@ -44,12 +58,28 @@
   langid = {english}
 }
 
+@article{hyman2014LiquidLiquid,
+  title = {Liquid-{{Liquid Phase Separation}} in {{Biology}}},
+  author = {Hyman, Anthony A. and Weber, Christoph A. and J{\"u}licher, Frank},
+  year = {2014},
+  month = oct,
+  journal = {Annual Review of Cell and Developmental Biology},
+  volume = {30},
+  number = {1},
+  pages = {39--58},
+  publisher = {Annual Reviews},
+  issn = {1081-0706, 1530-8995},
+  doi = {10.1146/annurev-cellbio-100913-013325},
+  urldate = {2024-04-12},
+  langid = {english}
+}
+
 @article{jacobs2023Theory,
   title = {Theory and {{Simulation}} of {{Multiphase Coexistence}} in {{Biomolecular Mixtures}}},
   author = {Jacobs, William M.},
   year = {2023},
   month = jun,
-  journal = {J. Chem. Theory Comput.},
+  journal = {Journal of Chemical Theory and Computation},
   volume = {19},
   number = {12},
   pages = {3429--3445},
@@ -192,14 +222,6 @@
   langid = {english}
 }
 
-@misc{Polymera,
-  title = {Polymer {{Property Predictor}} and {{Database}}},
-  urldate = {2024-12-02},
-  year = 2024,
-  howpublished = {https://pppdb.uchicago.edu/},
-  note = {Accessed: 2024-12-10}
-}
-
 @misc{qiang2024Scaling,
   title = {Scaling of Phase Count in Multicomponent Liquids},
   author = {Qiang, Yicheng and Luo, Chengjie and Zwicker, David},
@@ -218,7 +240,7 @@
   author = {Schmid, F.},
   year = {1998},
   month = sep,
-  journal = {J. Phys.: Condens. Matter},
+  journal = {Journal of Physics: Condensed Matter},
   volume = {10},
   number = {37},
   pages = {8105},
@@ -233,7 +255,7 @@
   author = {Shrinivas, Krishna and Brenner, Michael P.},
   year = {2021},
   month = nov,
-  journal = {Proc. Natl. Acad. Sci. U.S.A.},
+  journal = {Proceedings of the National Academy of Sciences},
   volume = {118},
   number = {45},
   pages = {e2108551118},
@@ -260,7 +282,7 @@
   author = {Sundman, Bo and Kattner, Ursula R. and Palumbo, Mauro and Fries, Suzana G.},
   year = {2015},
   month = dec,
-  journal = {Integr Mater Manuf Innov},
[…]
