+++ docs/paper/paper.bib
@@ -139,7 +139,7 @@
   shorttitle = {Computational Thermodynamics},
   author = {Lukas, Hans Leo and Fries, Suzana G. and Sundman, Bo},
   year = {2007},
-  publisher = {Cambridge Univ. Press},
+  publisher = {Cambridge University Press},
   address = {Cambridge},
   isbn = {978-0-521-86811-2},
   langid = {english}
