### 2025-02-18 #7388 @qiangyicheng
@srmnitc Thank you for the comments! We have just revised the manuscripts accordingly. We now use the full journal names consistently.

Thank you very much for all of your efforts on the package and the manuscript! @srmnitc @mastricker @SunyongKwon
